+++ code/src/main/kotlin/domain/Rates.kt
@@ -1,0 +1,8 @@
+package domain
+
+data class Rates(
+    val usd: Float,
+    val czk: Float,
+    val pln: Float,
+    val rub: Float
+)
